// frontend/state/interfaces/interfaces.ts
export interface IProperties {
  bathrooms: number;
  rooms: number;
  coveredArea: number;
  title: string;
  price: number;
  totalArea: number;
  id: string;
  slug: string; // Generated by https://quicktype.io
  operation: string;
  images: string[];
}

export interface IProperty {
  id: string;
  title: string;
  slug: string;
  price: number;
  rooms: number;
  bathrooms: number;
  totalArea: number;
  coveredArea: number;
  description: string;
  address: string;
  type: string;
  operation: string;
  bedrooms: null;
  antiquity: number;
  images: Image[];
  metaKeywords: string;
  metaDescription: string;
}

export interface Image {
  id: number;
  url: string;
}

export interface IAuthResponse {
  email: string;
  roles: ["user" | "admin"];
  token: string;
}